refactor: move FetchLatestEvent inside util package
