--- pkg/analyzer/events.go
@@ -1,50 +1,0 @@
-/*
-Copyright 2023 The K8sGPT Authors.
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-    http://www.apache.org/licenses/LICENSE-2.0
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package analyzer
-
-import (
-	"context"
-
-	"github.com/k8sgpt-ai/k8sgpt/pkg/kubernetes"
-	v1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-)
-
-func FetchLatestEvent(ctx context.Context, kubernetesClient *kubernetes.Client, namespace string, name string) (*v1.Event, error) {
-
-	// get the list of events
-	events, err := kubernetesClient.GetClient().CoreV1().Events(namespace).List(ctx,
-		metav1.ListOptions{
-			FieldSelector: "involvedObject.name=" + name,
-		})
-
-	if err != nil {
-		return nil, err
-	}
-	// find most recent event
-	var latestEvent *v1.Event
-	for _, event := range events.Items {
-		if latestEvent == nil {
-			// this is required, as a pointer to a loop variable would always yield the latest value in the range
-			e := event
-			latestEvent = &e
-		}
-		if event.LastTimestamp.After(latestEvent.LastTimestamp.Time) {
-			// this is required, as a pointer to a loop variable would always yield the latest value in the range
-			e := event
-			latestEvent = &e
-		}
-	}
-	return latestEvent, nil
-}
